+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Next Release
 Features:
 
 * Support `[inlet~]` object's control inlet and outlet
+* DPF: support Event type using a button for ImGui
 
 Bugs: